move boilerplate code to PaidContentPresenterBase
